Add files via upload
A responsive website, built with Webpack, written with the BEM syntax using Sass. The part of it, containing data, depending on the screen size, dynamically renders a table or a list of the above-mentioned data.
This shows how to use the .machMedia () method instead of the older method - .resize (). In addition, it contains the HTML tag <picture> instead of <img>, which allows you to indicate the use of the photo in the appropriate size depending on the screen size.